update graphics.py
Переписал вывод графика через класс.
Исправил метод подсчёта суммы.
Теперь можно задать операцию вручную.
