+++ config.h
@@ -1,0 +1,4 @@
+#pragma once
+#include <string>
+
+inline const std::string path_data = "D:\\Documents\\GitHub\\LightningTree\\LightningTree_data";
